+++ packages/terra-table/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added clinical-lowlight-theme theming files
+
+4.1.0 - (February 4, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 4.0.0 - (January 17, 2020)
 ------------------